+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Implement custom hook for inventory management with CRUD operations
+- Add custom hook for managing filters with search and sort functionality
 
 ## [0.6.0] - 2026-03-08