+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implement custom hook for inventory management with CRUD operations
 - Add custom hook for managing filters with search and sort functionality
+- Implement custom hook for managing shopping list with add, remove, and update functionalities
 
 ## [0.6.0] - 2026-03-08